backend/requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Mail==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
Werkzeug==3.1.3
